String extraction + reformat
This version of string extraction would handle well over 95% of the strings, what is left to do here is adding handling for conditional string reformat, the time that might spend on this feature plus how few instances conditional string format is being used does not justify completing this handling.

Oh, also the ambitious imp's ;) face is disrupting the reformat, won't add exception for the same reason.

Also, quite a few strings that shouldn't be inject&formatted are included, notably assets reading, need to rollback the changes here.

Next commmit would be reverting some changes and manually edit string.json so it would be ready for implementation, this commit contains quite a few error that prevent compiling which is intentional.